Move the function create_nncf_test_case
